split curl headers
